Avoid deprecated JAX behavior
This currently raises a deprecation warning because it is implicitly converting a length-1 array to a scalar. This is a deprecated path, and will raise an error after https://github.com/google/jax/pull/19181.

The easiest fix for now is to cast to `np.array`, because `rates` is also a NumPy array. Such scalar casts are also deprecated in NumPy, but will not error until NumPy 2.0, so the fix is safe for the time being.

A more long-term change requires understanding the range of array shapes expected to be handled by this code path, and would be better addressed by someone more familiar with the codebase.

PiperOrigin-RevId: 607489526